backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-multipart==0.0.6
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
pydantic-settings==2.1.0
pydantic[email]==2.4.2
aiosqlite==0.22.1



